# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: PongCode API
  version: 2.0.0
  description: >
    PongCode 前后端分离后的现有 Flask 接口契约。该文件只描述当前接口，
    不改变服务端业务、权限、数据库或文件上传行为。
servers:
  - url: /
...
paths:
...
  /api/organizations/{orgId}/projects:
    parameters:
      - $ref: "#/components/parameters/OrgId"
    post:
      operationId: createProject
      requestBody: *json-body
      responses: *created-response
...
  /api/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: getProject
      responses: *json-response
    put:
      operationId: updateProject
      requestBody: *json-body
      responses: *json-response
    delete:
      operationId: deleteProject
      responses: *json-response
  /api/projects/{projectId}/feishu-bot:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: getProjectFeishuBot
      responses:
        "200":
          description: 飞书机器人脱敏状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeishuBotStatus"
    put:
      operationId: updateProjectFeishuBot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                webhook_url:
                  type: string
                secret:
                  type: string
      responses:
        "200":
          description: 飞书机器人脱敏状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeishuBotStatus"
    delete:
      operationId: deleteProjectFeishuBot
      responses: *json-response
  /api/projects/{projectId}/feishu-bot/test:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      operationId: testProjectFeishuBot
      responses: *json-response
...
components:
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: integer
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    FeishuBotStatus:
      type: object
      required:
        - enabled
        - secret_configured
      properties:
        enabled:
          type: boolean
        webhook_masked:
          type:
            - string
            - "null"
        secret_configured:
          type: boolean
...